FIX: Rydder vekk muligheten til å løse 5090, som ikke er blitt utledet siden mars 2021
